## 🚀 Overview
This PR significantly improves the performance, reliability, and maintainability of the `readiluks` test suite. By optimizing LUKS container creation and strengthening teardown and error handling logic, the test suite runs up to **80% faster** while increasing test coverage and confidence.

---

## 🔧 Key Changes

### 1. **Performance Optimization of `setup_luks()`**
- Reduced cryptographic workload for LUKS containers in the test environment.
- Switched to `argon2id` with minimal settings to speed up container creation:
  - `--pbkdf-memory 512`
  - `--pbkdf-parallel 4`
  - `--pbkdf-force-iterations 4`
- Reduced `setup_luks()` runtime from **46 seconds** to **~3 seconds**, cutting **setup_lvm()** and **format_filesystem()** times as well.
- ✅ Realistic LUKS containers are still created, maintaining test fidelity.

### 2. **Expanded Test Coverage for `setup_lvm()`**
- Added comprehensive failure condition tests, covering:
  - `dmsetup info` missing major/minor numbers.
  - Existing logical volumes.
  - `vgcreate`, `lvcreate`, `vgchange -ay`, and `lvchange -ay` failures.
- Introduced stricter test isolation:
  - Preserved and restored `PATH`.
  - Cleaned up mock binaries and binaries like `dmsetup`.

### 3. **Enhanced Teardown Validation in `teardown_device()`**
- Added tests for:
  - `blkid` detection of leftover signatures (failure validation).
  - `cryptsetup erase` failure with `isLuks` returning true.
  - `wait_for_removal` timeouts on persistent conditions.
- These test cases ensure critical teardown error handling logic is exercised.

### 4. **Refactoring for Cleanup and Maintainability**
- `setup_lvm()` now cleans up PV, VG, and LV if any step fails.
- Introduced `wait_for_removal()` to unify wait loops:
  - Used across teardown tasks (unmounting, removing LVM components, etc.).
- Result: reduced code duplication, clearer error handling, and better resource cleanup.

### 5. **Test Coverage Runner Enhancements**
- Faster test runs by eliminating redundant coverage collection.
- Added `print_timing_line()` and `calculate_coverage()` for clearer reports.
- Removed `bc` dependency in favor of `awk` for portability.
- Coverage reports now show two decimal precision, aligned output, and missing lines.

---

## ⏱️ Performance Improvements
| Test Stage               | **Before** | **After**  | **Improvement**      |
|--------------------------|-----------:|-----------:|----------------------|
| `register_test_device`   |      2.874s |      2.855s | ⬇️ 0.7%              |
| `setup_luks`             |     46.106s |      3.388s | 🚀 92.6% faster      |
| `setup_lvm`              |    140.436s |     20.057s | 🚀 85.7% faster      |
| `format_filesystem`      |    100.476s |     23.437s | 🚀 76.7% faster      |
| `teardown_device`        |     87.480s |     26.052s | 🚀 70.2% faster      |
| **Total Runtime**        |    377.408s |     75.826s | 🚀 79.9% faster      |

---

## ✅ Benefits
- ⚡ **80% faster** test suite runtime.
- 🔒 **Improved cleanup and error handling** in device lifecycle tests.
- 📈 **Increased test coverage**, hitting previously uncovered teardown and setup failure paths.
- 🔧 **More maintainable test code**, reducing duplication and complexity.
- 🚀 Faster iteration cycles for development and CI/CD.

---

## 📌 What's Next
While further performance could be achieved with `GNU parallel` or `BATS` parallelization, the current improvements meet the project's needs. Focus will now shift to core logic development for `readiluks` and preparing for broader feature rollout.

---

## 🙌 Request for Review
Looking forward to feedback on:
- Further test isolation or coverage suggestions.
- Any concerns regarding the balance between performance tuning and test realism.

---

- **This commit refactors the test coverage runner to improve efficiency, output clarity, and eliminate redundant runs.**
- **``` refactor: add cleanup to setup_lvm error handling and consolidate wait loops**
- **``` test(setup_lvm): add comprehensive failure condition tests for setup_lvm()**
- **``` test(teardown_device): add tests for final validation and error handling paths**
- **feat(test_device_fixture): optimize LUKS container creation for faster test execution**